requirements.txt
psycopg2==2.9.10
pymonetdb==1.8.3
PyYAML==6.0.2
kagglehub==0.3.6
pandas==2.2.3
